requirements.txt
quart
sqlalchemy
uvicorn
pillow
aiofiles
argon2-cffi
